uint unique ID support

+++ Assets/QuickSheet/Example/Data/Runtime/ExcelExampleData.cs
@@ -9,16 +9,16 @@ using System.Collections;
 public class ExcelExampleData
 {
   [SerializeField]
-  int id;
-  public int Id { get {return id; } set { this.id = value;} }
+  uint id;
+  public uint Id { get {return id; } set { this.id = value;} }
   
   [SerializeField]
   string name;
   public string Name { get {return name; } set { this.name = value;} }
   
   [SerializeField]
-  float strength;
-  public float Strength { get {return strength; } set { this.strength = value;} }
+  long strength;
+  public long Strength { get {return strength; } set { this.strength = value;} }
   
   [SerializeField]
   Difficulty difficulty;
